Add Granta MI 2024 R1 integration tests (#189)

+++ tests/conftest.py
@@ -27,7 +27,8 @@ import warnings
 import pytest
 
 from ansys.grantami.jobqueue import Connection, JobQueueApiClient
-from common import FOLDER_NAME, delete_record, generate_now, get_granta_mi_version
+from ansys.grantami.jobqueue._connection import MINIMUM_GRANTA_MI_VERSION
+from common import FOLDER_NAME, delete_record, generate_now
 
 
 @pytest.fixture(scope="session")
@@ -56,10 +57,14 @@ def password_read_permissions():
 
 
 @pytest.fixture(scope="session")
-def job_queue_api_client(sl_url, admin_username, admin_password):
+def job_queue_api_client(sl_url, admin_username, admin_password, mi_version):
     """
     Fixture providing a real ApiClient to run integration tests against an instance of Granta MI
     Server API.
+
+    If client cannot be created because an unsupported Granta MI version is under test, instead yield
+    None and skip teardown. Tests that rely on a client being successfully generated should be
+    skipped via the 'mi_version' argument to the integration mark.
     """
     if all([admin_username, admin_password]):
         connection = Connection(sl_url).with_credentials(admin_username, admin_password)
@@ -67,18 +72,29 @@ def job_queue_api_client(sl_url, admin_username, admin_password):
         connection = Connection(sl_url).with_autologon()
     else:
         raise ValueError("Specify both or neither of TEST_ADMIN_USER and TEST_ADMIN_PASS.")
-    client: JobQueueApiClient = connection.connect()
-    clear_job_queue(client)
-    delete_record(
-        client=client,
-        name=FOLDER_NAME,
-    )
+
+    skip_teardown = False
+    try:
+        client: JobQueueApiClient = connection.connect()
+    except ConnectionError as e:
+        if mi_version < MINIMUM_GRANTA_MI_VERSION:
+            client = None
+            skip_teardown = True
+        else:
+            raise e
+    else:
+        clear_job_queue(client)
+        delete_record(
+            client=client,
+            name=FOLDER_NAME,
+        )
     yield client
-    clear_job_queue(client)
-    delete_record(
-        client=client,
-        name=FOLDER_NAME,
-    )
+    if not skip_teardown:
+        clear_job_queue(client)
+        delete_record(
+            client=client,
+            name=FOLDER_NAME,
+        )
 
 
 @pytest.fixture(scope="function")
@@ -117,24 +133,20 @@ def clear_job_queue(client: JobQueueApiClient):
         warnings.warn(f"Cleanup failed because of {e}\n continuing tests for now")
 
 
+def pytest_addoption(parser):
+    parser.addoption("--mi-version", action="store", default=None)
+
+
 @pytest.fixture(scope="session")
-def mi_version(request) -> tuple[int, int] | None:
-    """The version of MI referenced by the test url.
-
-    Returns
-    -------
-    tuple[int, int] | None
-        A 2-tuple containing the (MAJOR, MINOR) Granta MI release version, or None if a test URL is not available.
-
-    Notes
-    -----
-    This fixture returns None if the ``sl_url`` variable is not available. This is typically because the tests are
-    running in CI and the TEST_SL_URL environment variable was not populated.
-    """
-    if os.getenv("CI") and not os.getenv("TEST_SL_URL"):
+def mi_version(request):
+    mi_version: str = request.config.getoption("--mi-version")
+    if not mi_version:
         return None
-    client = request.getfixturevalue("job_queue_api_client")
-    return get_granta_mi_version(client)
+    parsed_version = mi_version.split(".")
+    if len(parsed_version) != 2:
+        raise ValueError("--mi-version argument must be a MAJOR.MINOR version number")
+    version_number = tuple(int(e) for e in parsed_version)
+    return version_number
 
 
 @pytest.fixture(autouse=True)
@@ -154,10 +166,11 @@ def process_integration_marks(request, mi_version):
         # No integration marker anywhere in the stack
         return
     if mi_version is None:
-        # We didn't get an MI version
-        # Unlikely to occur, since if we didn't get an MI version we don't have a URL, so we can't run integration
-        # tests anyway
-        return
+        # We didn't get an MI version. If integration tests were requested, an MI version must be provided. Raise
+        # an exception to fail all tests.
+        raise ValueError(
+            "No Granta MI Version provided to pytest. Specify Granta MI version with --mi-version MAJOR.MINOR."
+        )
 
     # Process integration mark arguments
     mark: pytest.Mark = request.node.get_closest_marker("integration")
@@ -170,6 +183,6 @@ def process_integration_marks(request, mi_version):
     if not isinstance(allowed_versions, list):
         raise TypeError("mi_versions argument type must be of type 'list'")
     if mi_version not in allowed_versions:
-        formatted_version = ".".join(str(x) for x in mi_version)
+        formatted_version = ".".join(str(v) for v in mi_version)
         skip_message = f'Test skipped for Granta MI release version "{formatted_version}"'
         pytest.skip(skip_message)
